==> 00_split_dataset.ipynb <==
import os
import glob
import random
import numpy as np
import pandas as pd
from sklearn.model_selection import KFold

src_path = "G:/LiverCancer/original/"

file_paths = [file_path for fold in os.listdir(src_path) for file_path in glob.glob(os.path.join(src_path, fold, "*.jpg"))]

random.shuffle(file_paths)

kf=KFold(n_splits=10)

file_paths = np.array(file_paths)

for train_index, holdout_index in kf.split(file_paths):

    train_file_paths = file_paths[train_index]
    holdout_file_paths = file_paths[holdout_index]
    
    df_train = pd.DataFrame(train_file_paths, columns=["file_path"])
    df_test = pd.DataFrame(holdout_file_paths, columns=["file_path"])
    
    dst_path = "G:/LiverCancer/resources/split_rule/"+ str(count) + "/"
    
    os.makedirs(dst_path, exist_ok=True)
        
    df_train.to_csv(dst_path + "train.csv" , index=False)
    df_test.to_csv(dst_path + "holdout.csv" , index=False)
