==> 07_resize_masked_biopsy.ipynb <==
import cv2
import glob
import ntpath

resize_masked_biopsys()

def resize_masked_biopsys():

    for grade_no in range(0, 5):
        
        src_path = "G:/LiverCancer/masked_biopsy/" + "grade" + str(grade_no) + "/"
        src_files = glob.glob(src_path + "*.jpg")
    
        for src_file in src_files:

            print(src_file)
            
            resize_masked_biopsy(src_file)

def resize_masked_biopsy(src_file):

    dst_path = "G:/LiverCancer/masked_biopsy_resized/"

    if "grade0" in src_file: sample_label = 0
    if "grade1" in src_file: sample_label = 1
    if "grade2" in src_file: sample_label = 2
    if "grade3" in src_file: sample_label = 3
    if "grade4" in src_file: sample_label = 4

    sample_image = cv2.imread(src_file)
    sample_image = cv2.resize(sample_image, (256,256)) 
    
    cv2.imwrite(dst_path + "grade" + str(sample_label) + "/" + ntpath.basename(src_file), sample_image)
